perf(dspark): drop the redundant fp32 cast on the base-logits slice
Both DSpark block samplers wrote `base_logits[:, k].float() + bias`. The
bias is already fp32, so torch promotes the bf16 slice to fp32 before the
add anyway -- the explicit cast just materializes a second [B, V] fp32
copy (42MB at B=64, V=163840) and reads it back, once per block position.

bf16 -> fp32 is exact and the promotion produces the same fp32 addends, so
the sum is bit-identical; this only removes work.

+++ atom/models/deepseek_v4_dspark.py
@@ -992,7 +992,10 @@ class _DSparkInner(nn.Module):
         markov_embeds = []
         for k in range(T):
             bias, m_embed = last.markov_head(out_ids[:, k])  # [B, V], [B, r]
-            logits_k = base_logits[:, k].float() + bias
+            # No .float() on the slice: bf16 + fp32 already promotes to fp32
+            # before the add, so casting first only materializes the [B, V]
+            # slice a second time for a bit-identical sum.
+            logits_k = base_logits[:, k] + bias
             out_ids[:, k + 1] = logits_k.argmax(
                 dim=-1
             )  # greedy (temp handled upstream)
